Migrates pull.sensor to new plugin layout
