+++ tests/plugins/pull/sensor/pyaudio_mock.py
@@ -1,23 +1,3 @@
-"""Some mocking stuff for packages that are not available on every platform (just for testing)."""
-
-import random
-
-
-class DHTMock:
-    """Mocks a DHT device (only available on rpi's)."""
-    DHT22 = "dht22"
-    DHT11 = "dht11"
-    AM2302 = "am2302"
-
-    @staticmethod
-    def read_retry(sensor, pin):  # pylint: disable=unused-argument
-        """Read the sensor values (humidity, temperature)."""
-        return (
-            round(random.uniform(1, 100), 2),
-            round(random.uniform(8, 36), 2)
-        )  # pragma: no cover
-
-
 class PyAudioMock:  # pragma: no cover
     """Mocks the pyaudio package (only available with audio devices)."""
     class _StreamMock:
